Fix imports with new directory structure
