+++ src/team-page/members/index.ts
@@ -1,0 +1,4 @@
+
+import { improveMemberCards } from "./content";
+
+improveMemberCards();
